% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{AdaptiveMetropolis}
\alias{AdaptiveMetropolis}
\alias{AdaptiveMetropolis-package}
\title{An R package implementing a series of adaptive metropolis hastings Markov
chain algorithms.}
\description{
Written originally as a project to satisfy the assesment requirements of the
Statistical Computing module from the Oxford-Warwick Statistics Programme
(OxWaSP) EPSRC Center for Doctoral Traning.
}

