% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AM}
\alias{AM}
\title{Adaptive Metropolis Sampler}
\usage{
AM(target, n, init.state, init.cov, burn.in = 1000)
}
\arguments{
\item{target}{A function that evaluates the unscaled densities of the target
distribution.}

\item{n}{The number of samples to produce.}

\item{init.state}{The initial state of the sampler.}

\item{init.cov}{The covariance of the sampler during burn in.}

\item{burn.in}{(Optional) Number of iterations to burn in for. Default: 1000}
}
\value{
\code{n} samples from the target distribution.
}
\description{
An implementation of the Adaptive Metropolis algorithm as proposed by Haario
et al (2001).
The covariance matrix of the proposal distribution is adapted with current
knowledge of the target distrubution.
}
\details{
The proposal distribution is a multivariate normal distribution of the
appropriate dimension. The covariance of this distribution is computed from
previous samples.

The dimension is detected using the length of the \code{init.state}
parameter.

Adaptive samplers do not attempt to solve the problem of multi modal
distributions. If the target distribution is known to have multiple modes,
particually if they are far apart, more advanced techniques are recomended
such as nesting adaptive Metropolis within parallel tempering.
}
\examples{
# 1D Normal Distribution
data <- AM(function(x) {dnorm(x, 0,100)}, 5000, 0, 1)
plot(data)

# 2D Uniform Distribution
data <- AM(function(x) {dunif(x, c(0,0), c(100,100))}, 5000, c(0,0), diag(2))
plot(data)
}
\author{
Tom Jin <sjin@stats.ox.ac.uk>
}
\references{
Haario, H., Saksman, E., Tamminen, J. (2001)
\emph{An Adaptive Metropolis Algorithm.} Bernoulli, Vol. 7, No. 2.
}

