% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AMWG}
\alias{AMWG}
\title{Adaptive Metropolis-Within-Gibbs}
\usage{
AMWG(adapt, N, K)
}
\arguments{
\item{adapt}{whether to use adapted MH or normal MH algorithm}

\item{N}{number of iterations}

\item{K}{number of latent variables}
}
\value{
A list containing simulated parameters paths (theta) and acceptance ratio paths (acceptance).
}
\description{
Adaptive Metropolis-Hastings algorithm within the Gibbs algorithm, applied to a specific hierachical model.
}
\details{
Update parameter one by one using a normal proposal, accept it with probability depending on the simulated path.
}
\examples{
out_noadapt <- AMWG(adapt = 0, 1000, 100)
theta_noadapt <- out_noadapt[[1]]
out_adapt <- AMWG(adapt = 1, 1000, 100)
theta_adapt <- out_adapt[[1]]
plot(theta_noadapt[,1])
points(theta_adapt[,1],col=2)
}
\author{
Xiaoyu Lu
}
\references{
Roberts, G. O., Rosenthal, J. S. (2006), \emph{Examples of Adaptive MCMC}
}

