% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ASWAM}
\alias{ASWAM}
\title{Adaptive Scaling Within Adaptive Metropolis Sampler}
\usage{
ASWAM(target, n, init.state, init.cov)
}
\arguments{
\item{target}{The target distribution where the unscaled densities can be
evaluated.}

\item{n}{The number of samples to produce.}

\item{init.state}{The initial state of the sampler.}

\item{init.cov}{The initial covariance of the sampler.}
}
\value{
n samples from the target distribution.
}
\description{
Adaptive Scaling Within Adaptive Metropolis Sampler
}

