% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AM}
\alias{AM}
\title{Adaptive Metropolis Sampler}
\usage{
AM(target, n, init.state, init.cov)
}
\arguments{
\item{target}{The target distribution where the unscaled densities can be
evaluated.}

\item{n}{The number of samples to produce.}

\item{init.state}{The initial state of the sampler.}

\item{init.cov}{The initial covariance of the sampler.}
}
\value{
n samples from the target distribution.
}
\description{
Adaptive Metropolis Sampler
}
\examples{
data <- AM(function(x) {dunif(x, 0,100)}, 1000, 0, 1)
plot(data)
data <- AM(function(x) {dunif(x, c(0,0), c(100,100))}, 1000, c(0,0), diag(2))
plot(data)
}

